Why Unzipped Hover
Created components for dropping down the unzipped hover options and applied filters as per requirements. Added some checks for the submenus to get open on demand and also fixed some issues related the dynamic dropdown. Added appropriate icons needed for the links and created schema for the sub-menu items as well. Added styling as per the figma design.